calendar boundaries cleaned up
