+++ cmd/root.go
@@ -16,6 +16,7 @@ package cmd
 
 import (
 	"fmt"
+	"flag"
 	"os"
 
 	homedir "github.com/mitchellh/go-homedir"
@@ -44,6 +45,7 @@ to quickly create a Cobra application.`,
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
+
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
@@ -56,8 +58,9 @@ func init() {
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.blync-studio-light.yaml)")
 	rootCmd.PersistentFlags().IntP("device", "d", 0, "Device index for light to interface with")
 	//nolint:errcheck
-	viper.BindPFlag("device", rootCmd.Flags().Lookup("device"))
+	viper.BindPFlag("device", rootCmd.PersistentFlags().Lookup("device"))
 	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	flag.Parse()
 }
 
 // initConfig reads in config file and ENV variables if set.
