+++ config/calendar.go
@@ -3,4 +3,5 @@ package config
 
 type GoogleCalendarConfiguration struct {
 	CalendarId string
+	Email string
 }
